More error handling fixes

+++ reflector/AbstractReflectSession.h
@@ -527,7 +527,7 @@ protected:
    MUSCLE_NODISCARD uint64 GetLastByteOutputTimeStamp() const {return _lastByteOutputAt;}
 
 private:
-   virtual void TallySubscriberTablesInfo(uint32 & retNumCachedSubscriberTables, uint32 & tallyNumNodes, uint32 & tallyNumNodeBytes) const;  // yes, this virtual method is intentionally private!
+   virtual void TallySubscriberTablesInfo(uint32 & retNumCachedSubscriberTables, uint64 & tallyNumNodes, uint64 & tallyNumNodeBytes) const;  // yes, this virtual method is intentionally private!
 
    void SetPolicyAux(AbstractSessionIOPolicyRef & setRef, uint32 & setChunk, const AbstractSessionIOPolicyRef & newRef, bool isInput);
    void PlanForReconnect();
@@ -550,7 +550,7 @@ private:
    IPAddressAndPort _asyncConnectDest;
    bool _reconnectViaTCP;            // only valid when _asyncConnectDest is set
    AbstractMessageIOGatewayRef _gateway;
-   uint64 _lastByteOutputAt;         // timestamp of last time we output a byt
+   uint64 _lastByteOutputAt;         // timestamp of last time we output a byte
    uint64 _pendingLastByteOutputAt;  // same as _lastByteOutputAt but 0 if we don't currently want to output anything
    uint32 _lastReportedQueueSize;    // used by ReflectServer.cpp to warn about growing/socket-free queues
    AbstractSessionIOPolicyRef _inputPolicyRef;
